Rename `borrow::internal` to `borrow::private`
